chore: abstract mobile deeplink handling
