+++ packages/appkit-utils/src/MobileWallet.ts
@@ -1,3 +1,11 @@
+import { ConstantsUtil } from '@reown/appkit-common'
+import {
+  ChainController,
+  ConnectorController,
+  RouterController,
+  type WcWallet
+} from '@reown/appkit-core'
+
 export const MobileWalletUtil = {
   /**
    * Handles mobile wallet redirection for wallets that have Universal Links.
@@ -5,20 +13,35 @@ export const MobileWalletUtil = {
    * @param {Object} properties - The properties object.
    * @param {string} properties.name - The name of the wallet.
    */
-  handleMobileWalletRedirection(properties: { name: string }): void {
-    const href = window.location.href
-    const encodedHref = encodeURIComponent(href)
+  handleMobileWalletRedirection(wallet: WcWallet): void {
+    const connector = ConnectorController.getConnector(wallet.id, wallet.rdns)
 
-    if (properties?.name === 'Phantom' && !('phantom' in window)) {
-      const protocol = href.startsWith('https') ? 'https' : 'http'
-      const host = href.split('/')[2]
-      const encodedRef = encodeURIComponent(`${protocol}://${host}`)
+    if (ChainController.state.activeChain === ConstantsUtil.CHAIN.SOLANA) {
+      /**
+       * Universal Links requires explicit user interaction to open the wallet app.
+       * Previously we've been calling this with the life-cycle methods in the Solana clients by listening the SELECT_WALLET event of EventController.
+       * But this breaks the UL functionality for some wallets like Phantom.
+       */
+      const href = window.location.href
+      const encodedHref = encodeURIComponent(href)
+      const name = connector?.name || wallet?.name || ''
+      if (name === 'Phantom' && !('phantom' in window)) {
+        const protocol = href.startsWith('https') ? 'https' : 'http'
+        const host = href.split('/')[2]
+        const encodedRef = encodeURIComponent(`${protocol}://${host}`)
 
-      window.location.href = `https://phantom.app/ul/browse/${encodedHref}?ref=${encodedRef}`
+        window.location.href = `https://phantom.app/ul/browse/${encodedHref}?ref=${encodedRef}`
+      }
+
+      if (name === 'Coinbase Wallet' && !('coinbaseSolana' in window)) {
+        window.location.href = `https://go.cb-w.com/dapp?cb_url=${encodedHref}`
+      }
     }
 
-    if (properties?.name === 'Coinbase Wallet' && !('coinbaseSolana' in window)) {
-      window.location.href = `https://go.cb-w.com/dapp?cb_url=${encodedHref}`
+    if (connector) {
+      RouterController.push('ConnectingExternal', { connector })
+    } else {
+      RouterController.push('ConnectingWalletConnect', { wallet })
     }
   }
 }
