+++ packages/scaffold-ui/src/partials/w3m-all-wallets-search/index.ts
@@ -1,14 +1,8 @@
 import { LitElement, html } from 'lit'
 import { property, state } from 'lit/decorators.js'
 
-import { ConstantsUtil } from '@reown/appkit-common'
 import type { BadgeType, WcWallet } from '@reown/appkit-core'
-import {
-  ApiController,
-  ChainController,
-  ConnectorController,
-  RouterController
-} from '@reown/appkit-core'
+import { ApiController } from '@reown/appkit-core'
 import { customElement } from '@reown/appkit-ui'
 import { MobileWalletUtil } from '@reown/appkit-utils'
 
@@ -100,24 +94,7 @@ export class W3mAllWalletsSearch extends LitElement {
   }
 
   private onConnectWallet(wallet: WcWallet) {
-    const connector = ConnectorController.getConnector(wallet.id, wallet.rdns)
-
-    if (ChainController.state.activeChain === ConstantsUtil.CHAIN.SOLANA) {
-      /**
-       * Universal Links requires explicit user interaction to open the wallet app.
-       * Previously we've been calling this with the life-cycle methods in the Solana clients by listening the SELECT_WALLET event of EventController.
-       * But this breaks the UL functionality for some wallets like Phantom.
-       */
-      MobileWalletUtil.handleMobileWalletRedirection({
-        name: connector?.name || wallet?.name || ''
-      })
-    }
-
-    if (connector) {
-      RouterController.push('ConnectingExternal', { connector })
-    } else {
-      RouterController.push('ConnectingWalletConnect', { wallet })
-    }
+    MobileWalletUtil.handleMobileWalletRedirection(wallet)
   }
 }
